refactor(view): rename package from input to view
